+++ src/Encamina.Enmarcha.AI.QuestionsAnswering.Azure/QuestionAnsweringService.cs
@@ -5,8 +5,6 @@ using CommunityToolkit.Diagnostics;
 
 using Encamina.Enmarcha.AI.QuestionsAnswering.Abstractions;
 
-using Microsoft.Extensions.Options;
-
 namespace Encamina.Enmarcha.AI.QuestionsAnswering.Azure;
 
 /// <summary>
@@ -65,7 +63,7 @@ internal class QuestionAnsweringService : CognitiveServiceBase<QuestionAnswering
         };
     }
 
-    private static QueryFilters BuildQueryFilters(IQuestionRequestOptions requestOptions)
+    private static QueryFilters? BuildQueryFilters(IQuestionRequestOptions requestOptions)
     {
         if (requestOptions == null)
         {
@@ -88,7 +86,7 @@ internal class QuestionAnsweringService : CognitiveServiceBase<QuestionAnswering
         return queryFilter.MetadataFilter != null || queryFilter.SourceFilter.Any() ? queryFilter : null;
     }
 
-    private static MetadataFilter BuildMetadataFilter(MetadataOptions metadataOptions)
+    private static MetadataFilter? BuildMetadataFilter(MetadataOptions metadataOptions)
     {
         if (metadataOptions == null || (!metadataOptions.Metadata?.Any() ?? true))
         {